Update basevalidators.py
Pandas will send timedelta index and series types with a dtype of "m". Timedeltas now only "work" because they are copied over as though they were plain integer arrays. The dtype information gets lost because only "M" is being caught and processed explicitly. By the time this is processed for clientside consumption, the raw integer values are not always obvious to the user as being the original specified values (such as when millis get translated to nano) so the fact that it currently "works" is a loose term. It doesn't fail explicitly is more correct. 

The underlying pandas index and series already have to_pytimedelta() ready for this use.